+++ src/explorepy/cli.py
@@ -12,8 +12,6 @@ import explorepy
 CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
 logger = logging.getLogger(__name__)
 
-default_bt_backend = explorepy.get_bt_interface()
-
 
 @click.group(context_settings=CONTEXT_SETTINGS, invoke_without_command=True)
 @click.option("--version", "-V", help="Print explorepy version", is_flag=True)
@@ -77,13 +75,22 @@ def acquire(name, address, duration):
 @click.option("-d", "--duration", type=int, help="Recording duration in seconds", metavar="<integer>")
 @click.option("--edf", 'file_type', flag_value='edf', help="Write in EDF file")
 @click.option("--csv", 'file_type', flag_value='csv', help="Write in csv file (default type)", default=True)
+@click.option("--imp-mode", is_flag=True, help="Enable impedance mode with live monitoring")
+@click.option("-nf", "--notch-freq", help="Notch frequency for impedance mode initialization", type=float, default=None)
 @verify_inputs
-def record_data(address, name, filename, overwrite, duration, file_type):
+def record_data(address, name, filename, overwrite, duration, file_type, imp_mode, notch_freq):
     """Record data from Explore to a file"""
     explore = explorepy.explore.Explore()
     explore.connect(mac_address=address, device_name=name)
-    explore.record_data(file_name=filename, file_type=file_type,
-                        do_overwrite=overwrite, duration=duration, block=True)
+    explore.record_data(
+        file_name=filename,
+        file_type=file_type,
+        do_overwrite=overwrite,
+        duration=duration,
+        imp_mode=imp_mode,
+        notch_freq=notch_freq,
+        block=True
+    )
 
 
 @cli.command()
@@ -153,17 +160,3 @@ def soft_reset(address, name):
     explore = explorepy.explore.Explore()
     explore.connect(mac_address=address, device_name=name)
     explore.reset_soft()
-
-
-@cli.command()
-@click.option("--address", "-a", type=str, help="Explore device's MAC address")
-@click.option("--name", "-n", type=str, help="Name of the device")
-@click.option("-m", "--channel-mask", type=str, required=True,
-              help="Channel mask, it should be a binary string containing 1 and 0, "
-                   "representing the mask (LSB is channel 1).")
-@verify_inputs
-def set_channels(address, name, channel_mask):
-    """Mask the channels of the Explore device"""
-    explore = explorepy.explore.Explore()
-    explore.connect(mac_address=address, device_name=name)
-    explore.set_channels(channel_mask)