+++ src/explorepy/packet.py
@@ -3,7 +3,6 @@
 import abc
 import binascii
 import logging
-import struct
 from enum import IntEnum
 
 import numba as nb
@@ -26,9 +25,10 @@ class PACKET_ID(IntEnum):
     DISCONNECT = 111
     # Info packet from BLE devices, applies to Explore Pro
     INFO_BLE = 98
+    INFO_HYP = 99
     # New info packet containing memory and board ID: this applies to all Explore+ systems
     INFO_V2 = 97
-    INFO = 99
+    INFO = 96
     EEG94 = 144
     EEG98 = 146
     EEG32 = 148
@@ -46,6 +46,7 @@ class PACKET_ID(IntEnum):
     CALIBINFO_USBC = 197
     TRIGGER_OUT = 177  # Trigger-out of Explore device
     TRIGGER_IN = 178  # Trigger-in to Explore device
+    VERSION_INFO = 199
 
 
 EXG_UNIT = 1e-6
@@ -64,7 +65,7 @@ class Packet(abc.ABC):
             time_offset (double): Time offset defined by parser. It will be the timestamp of the first packet when
                                     streaming in realtime. It will be zero while converting a binary file.
         """
-        self.timestamp = timestamp + time_offset
+        self.timestamp = timestamp
         self._convert(payload[:-4])
         self._check_fletcher(payload[-4:])
 
@@ -246,13 +247,7 @@ class EEG(Packet):
         If exg_fs is given, it returns time vector and data. If exg_fs is not given, it returns the timestamp of the
         packet alongside with the data
         """
-        if exg_fs:
-            n_sample = self.data.shape[1]
-            time_vector = np.linspace(self.timestamp,
-                                      self.timestamp + (n_sample - 1) / exg_fs,
-                                      n_sample)
-            return time_vector, self.data
-        return self.timestamp, self.data
+        return np.array([self.timestamp]), self.data
 
     def get_impedances(self):
         """get electrode impedances"""
@@ -508,65 +503,69 @@ class PushButtonMarker(EventMarker):
                                       np.uint16).newbyteorder("<"))[0]
 
 
-class SoftwareMarker(EventMarker):
-    """Software marker packet"""
-
-    def __init__(self, timestamp, payload, time_offset=0):
-        super().__init__(timestamp, payload, time_offset)
-        self._label_prefix = "sw_"
-
-    def _convert(self, bin_data):
-        self.code = np.frombuffer(bin_data,
-                                  dtype=np.dtype(
-                                      np.uint16).newbyteorder("<"))[0]
-
-    @staticmethod
-    def create(local_time, code):
-        """Create a software marker
-
-        Args:
-            local_time (double): Local time from LSL
-            code (int): Event marker code
-
-        Returns:
-            SoftwareMarker
-        """
-        return SoftwareMarker(
-            local_time * 10000,
-            payload=bytearray(struct.pack("<H", code) + b"\xaf\xbe\xad\xde"),
-        )
-
-
 class ExternalMarker(EventMarker):
     """External marker packet"""
 
-    def __init__(self, timestamp, payload, time_offset=0):
+    def __init__(self, timestamp, payload, name):
         super().__init__(timestamp, payload, 0)
-        self._label_prefix = "sw_"
+        self._label_prefix = "lsl_"
+        self.name = name
 
     def _convert(self, bin_data):
         self.code = bin_data[:15].decode('utf-8', errors='ignore')
 
     @staticmethod
-    def create(lsl_time, marker_string):
+    def create(lsl_time, marker_string, name):
         """Create a software marker
 
         Args:
             lsl_time (double): Local time from LSL
             marker_string (string): Event marker code
+            name (string): Marker stream name
 
         Returns:
             SoftwareMarker
         """
         if not isinstance(marker_string, str):
             raise ValueError("Marker label must be a string")
-        if len(marker_string) > 7 or len(marker_string) < 1:
+        if len(marker_string) > 20 or len(marker_string) < 1:
             raise ValueError(
-                "Marker label length must be between 1 and 7 characters")
+                "Marker label length must be between 1 and 20 characters")
         byte_array = bytes(marker_string, 'utf-8')
         return ExternalMarker(
             lsl_time,
             payload=bytearray(byte_array + b"\xaf\xbe\xad\xde"),
+            name=name
+        )
+
+
+class SoftwareMarker(ExternalMarker):
+    def __init__(self, timestamp, payload, name):
+        super().__init__(timestamp, payload, name)
+        self._label_prefix = 'sw_'
+
+    @staticmethod
+    def create(lsl_time, marker_string, name):
+        """Create a software marker
+
+        Args:
+            lsl_time (double): Local time from LSL
+            marker_string (string): Event marker code
+            name (string): Marker stream name
+
+        Returns:
+            SoftwareMarker
+        """
+        if not isinstance(marker_string, str):
+            raise ValueError("Marker label must be a string")
+        if len(marker_string) > 20 or len(marker_string) < 1:
+            raise ValueError(
+                "Marker label length must be between 1 and 20 characters")
+        byte_array = bytes(marker_string, 'utf-8')
+        return SoftwareMarker(
+            lsl_time,
+            payload=bytearray(byte_array + b"\xaf\xbe\xad\xde"),
+            name=name
         )
 
 
@@ -684,6 +683,10 @@ class DeviceInfoBLE(DeviceInfoV2):
         return as_dict
 
 
+class DeviceInfoHyp(DeviceInfoBLE):
+    pass
+
+
 class CommandRCV(Packet):
     """Command Status packet"""
 
@@ -782,6 +785,7 @@ PACKET_CLASS_DICT = {
     PACKET_ID.INFO: DeviceInfo,
     PACKET_ID.INFO_V2: DeviceInfoV2,
     PACKET_ID.INFO_BLE: DeviceInfoBLE,
+    PACKET_ID.INFO_HYP: DeviceInfoHyp,
     PACKET_ID.EEG94: EEG94,
     PACKET_ID.EEG98: EEG98,
     PACKET_ID.EEG99: EEG99,
@@ -799,4 +803,5 @@ PACKET_CLASS_DICT = {
     PACKET_ID.PUSHMARKER: PushButtonMarker,
     PACKET_ID.TRIGGER_IN: TriggerIn,
     PACKET_ID.TRIGGER_OUT: TriggerOut,
+    PACKET_ID.VERSION_INFO: VersionInfoPacket,
 }